Ios test cleanup (#673)
### This Commit:
Splits iOS test code into logical groups and adds comments to provide context for developers
