+++ ios/COVIDSafePathsTests/ExposureManagerUnitTests.swift
@@ -54,7 +54,7 @@ class MockENExposureInfo: ENExposureInfo {
 class MockENExposureDaySummary: ENExposureDaySummary {
   
   override var date: Date {
-    return halloween
+    return XCTestCase.halloween
   }
   
   var daySummaryHandler: (() -> ENExposureSummaryItem)?
@@ -288,7 +288,7 @@ class UNUserNotificationCenterMock: UserNotificationCenter {
 
 // MARK: == UNIT TESTS ==
 
-class ExposureManagerTests: XCTestCase {
+class ExposureManagerUnitTests: XCTestCase {
   
   func testCreatesSharedInstance() {
     ExposureManager.createSharedInstance()
@@ -340,7 +340,9 @@ class ExposureManagerTests: XCTestCase {
                registerNotificationExpectation,
                setExposureNotificationEnabledTrueExpectation], timeout: 1)
   }
-  
+
+  // When `awake` is called
+  // an `onEnabledStatusUpdated` notification is posted
   func testAwake() {
     
     let broadcastAuthorizationStateExpectation = self.expectation(description: "A notification is post with the current authorization and enabled stated")
@@ -354,150 +356,69 @@ class ExposureManagerTests: XCTestCase {
     exposureManager.awake()
     wait(for: [broadcastAuthorizationStateExpectation], timeout: 0)
   }
-  
-  func testEnabledtatus() {
-    
-    let mockENManager = ENManagerMock()
-    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
-    
-    mockENManager.exposureNotificationEnabledHandler = {
-      return true
-    }
-    XCTAssertEqual(exposureManager.enabledState, ExposureManager.EnabledState.enabled)
-    mockENManager.exposureNotificationEnabledHandler = {
-      return false
-    }
-    XCTAssertEqual(exposureManager.enabledState, ExposureManager.EnabledState.disabled)
-  }
-  
-  func testAuthorizationStatus() {
-    
-    let mockENManager = ENManagerMock()
-    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
-    
-    mockENManager.authorizationStatusHandler = {
-      return .authorized
-    }
-    XCTAssertEqual(exposureManager.authorizationState,
-                   ExposureManager.AuthorizationState.authorized)
-    mockENManager.authorizationStatusHandler = {
-      return .notAuthorized
-    }
-    XCTAssertEqual(exposureManager.authorizationState,
-                   ExposureManager.AuthorizationState.unauthorized)
-    mockENManager.authorizationStatusHandler = {
-      return .restricted
-    }
-    XCTAssertEqual(exposureManager.authorizationState,
-                   ExposureManager.AuthorizationState.unauthorized)
-    mockENManager.authorizationStatusHandler = {
-      return .unknown
-    }
-    XCTAssertEqual(exposureManager.authorizationState,
-                   ExposureManager.AuthorizationState.unauthorized)
-  }
-  
-  func testBluetoothStatus() {
-    
-    let mockENManager = ENManagerMock()
-    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
-    
-    mockENManager.exposureNotificationStatusHandler = {
-      return .bluetoothOff
-    }
-    XCTAssertFalse(exposureManager.isBluetoothEnabled)
-    mockENManager.exposureNotificationStatusHandler = {
-      return .active
-    }
-    XCTAssertTrue(exposureManager.isBluetoothEnabled)
-    mockENManager.exposureNotificationStatusHandler = {
-      return .disabled
-    }
-    XCTAssertTrue(exposureManager.isBluetoothEnabled)
-    mockENManager.exposureNotificationStatusHandler = {
-      return .paused
-    }
-    XCTAssertTrue(exposureManager.isBluetoothEnabled)
-    mockENManager.exposureNotificationStatusHandler = {
-      return .restricted
-    }
-    XCTAssertTrue(exposureManager.isBluetoothEnabled)
-    mockENManager.exposureNotificationStatusHandler = {
-      return .unknown
-    }
-    XCTAssertTrue(exposureManager.isBluetoothEnabled)
-  }
-  
+
+  // When calling `currentExposures` on an ExposureManager instance
+  // the call returns a string that can be serialized into an array of Exposures
   func testCurrentExposures() {
     let enManagerMock = ENManagerMock()
-    
-    enManagerMock.enDetectExposuresHandler = { configuration, diagnosisKeys, completionHandler in
-      let enExposureSummary = MockENExposureDetectionSummary()
-      enExposureSummary.matchedKeyCountHandler = {
-        return 1
-      }
-      enExposureSummary.attenuationDurationsHandler = {
-        return [900,0,0]
-      }
-      completionHandler(enExposureSummary, nil)
-      return Progress()
-    }
-    
-    enManagerMock.detectExposuresHandler = { configuration, completion in
-      completion(nil, GenericError.unknown)
-      return Progress()
-    }
-    
+
     let apiClientMock = APIClientMock { (request, requestType) -> (AnyObject) in
       return Result<String>.success(String.default) as AnyObject
-    }
-    apiClientMock.downloadRequestHander = { (request, requestType) in
-      if #available(iOS 13.7, *) {
-        return Result<DailySummariesConfiguration>.success(DailySummariesConfiguration.placeholder) as AnyObject
-      }
-      return Result<ExposureConfigurationV1>.success(ExposureConfigurationV1.placeholder) as AnyObject
     }
     
     let btSecureStorageMock = BTSecureStorageMock(notificationCenter: NotificationCenter())
     btSecureStorageMock.userStateHandler = {
       let userState = UserState()
       userState.exposures.append(Exposure(id: "1",
-                                          date: 0))
+                                          date: Date().posixRepresentation))
       return userState
     }
     let exposureManager = ExposureManager(exposureNotificationManager: enManagerMock,
                                           apiClient: apiClientMock,
                                           btSecureStorage: btSecureStorageMock)
-    let currentExposures = exposureManager.currentExposures
-    XCTAssertNoThrow(try! JSONDecoder().decode(Array<Exposure>.self, from: currentExposures.data(using: .utf8) ?? Data()))
+    let currentExposuresString = exposureManager.currentExposures
+    let currentExposures = try! JSONDecoder().decode(Array<Exposure>.self, from: currentExposuresString.data(using: .utf8) ?? Data())
+    XCTAssertEqual(currentExposures.count, 1)
   }
-  
+
+  // When `requestExposureNotificationAuthorization` is called
+  // the request is forwarded to the underlying manager
+  // and if the underlying manager successfully enables exposure notifications
+  // and an `onEnabledStatusUpdated` notification is posted
   func testEnableNotificationsSuccess() {
-    let setEnabled = true
     let setExposureNotificationEnabledExpectation = self.expectation(description: "Request the change of state to the underlying manager")
+    let broadcastAuthorizationStateExpectation = self.expectation(description: "A notification is post with the current authorization and enabled stated")
+
     let mockENManager = ENManagerMock()
+
     mockENManager.setExposureNotificationEnabledHandler = { enabled, completionHandler in
-      if enabled == setEnabled {
+      if enabled == true {
         setExposureNotificationEnabledExpectation.fulfill()
       }
       completionHandler(nil)
     }
-    let broadcastAuthorizationStateExpectation = self.expectation(description: "A notification is post with the current authorization and enabled stated")
+
     let notificationCenterMock = NotificationCenterMock()
+
     notificationCenterMock.postHandler = { notification in
       if notification.name == .AuthorizationStatusDidChange {
         broadcastAuthorizationStateExpectation.fulfill()
       }
     }
+
     let exposureManager = ExposureManager(exposureNotificationManager: mockENManager,
                                           notificationCenter: notificationCenterMock)
-    exposureManager.requestExposureNotificationAuthorization(enabled: setEnabled) { (error) in
+    exposureManager.requestExposureNotificationAuthorization(enabled: true) { (error) in
       XCTAssertNil(error, "There should be no error")
     }
     wait(for: [setExposureNotificationEnabledExpectation,
                broadcastAuthorizationStateExpectation], timeout: 0)
   }
-  
+
+  // When `requestExposureNotificationAuthorization` is called
+  // the request is forwarded to the underlying manager
+  // and if the underlying manager does not successfully enable exposure notifications
+  // and an `onEnabledStatusUpdated` notification is posted
   func testEnableNotificationsError() {
     let setExposureNotificationEnabledExpectation = self.expectation(description: "Request the change of state to the underlying manager")
     let mockENManager = ENManagerMock()
@@ -521,29 +442,17 @@ class ExposureManagerTests: XCTestCase {
     wait(for: [setExposureNotificationEnabledExpectation,
                broadcastAuthorizationStateExpectation], timeout: 0)
   }
-  
-  func testGetCurrentENPermissionsStatus() {
-    let mockENManager = ENManagerMock()
-    mockENManager.authorizationStatusHandler = {
-      return .authorized
-    }
-    mockENManager.exposureNotificationEnabledHandler = {
-      return true
-    }
-    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
-    exposureManager.getCurrentENPermissionsStatus { (authorized, enabled) in
-      XCTAssertEqual(authorized, ExposureManager.AuthorizationState.authorized.rawValue)
-      XCTAssertEqual(enabled, ExposureManager.EnabledState.enabled.rawValue)
-    }
-  }
-  
+
+  // When calling `notifyUserBlueToothOffIfNeeded`
+  // when bluetooth has been disabled
+  // a notification request is added with the appropriate title and body
   func testBluetoothNotificationOn() {
-    let addNotificatiionRequestExpectation = self.expectation(description: "A notification request is added with the proper title and body")
+    let addNotificationRequestExpectation = self.expectation(description: "A notification request is added with the proper title and body")
     let removeNotificationsExpectation = self.expectation(description: "when is not authorized and bluetooth is not off we just remove all delivered notifications")
-    
+
     let unUserNotificationCenterMock = UNUserNotificationCenterMock()
     unUserNotificationCenterMock.addHandler = { request, completionHandler in
-      addNotificatiionRequestExpectation.fulfill()
+      addNotificationRequestExpectation.fulfill()
       let content = request.content
       XCTAssertEqual(request.identifier, String.bluetoothNotificationIdentifier)
       XCTAssertEqual(content.title, String.bluetoothNotificationTitle.localized)
@@ -561,17 +470,15 @@ class ExposureManagerTests: XCTestCase {
     }
     let exposureManager = ExposureManager(exposureNotificationManager: mockENManager,
                                           userNotificationCenter: unUserNotificationCenterMock)
-    addNotificatiionRequestExpectation.isInverted = false
+    addNotificationRequestExpectation.isInverted = false
     removeNotificationsExpectation.isInverted = true
     exposureManager.notifyUserBlueToothOffIfNeeded()
-    wait(for: [addNotificatiionRequestExpectation, removeNotificationsExpectation], timeout: 0)
+    wait(for: [addNotificationRequestExpectation, removeNotificationsExpectation], timeout: 0)
   }
-  
-  func testRecentExposures() {
-    let withExposures = defaultStorage()
-    XCTAssertEqual(withExposures.userState.recentExposures.count, 1)
-  }
-  
+
+  // When calling `notifyUserBlueToothOffIfNeeded`
+  // when exposureNotificationStatus is not `bluetoothOff`
+  // notification is removed from UNUserNotificationCenter
   func testBluetoothNotificationOff() {
     let addNotificatiionRequestExpectation = self.expectation(description: "A notification request is added with the proper title and body")
     let removeNotificationsExpectation = self.expectation(description: "when is not authorized and bluetooth is not off we just remove all delivered notifications")
@@ -592,331 +499,6 @@ class ExposureManagerTests: XCTestCase {
     removeNotificationsExpectation.isInverted = false
     exposureManager.notifyUserBlueToothOffIfNeeded()
     wait(for: [addNotificatiionRequestExpectation, removeNotificationsExpectation], timeout: 0)
-  }
-  
-  func testFetchExposureKeysSuccess() {
-    let expectation = self.expectation(description: "a call is made to get the diagnosis keys")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    exposureManager.fetchExposureKeys { (keys, error) in
-      XCTAssertNil(error)
-      XCTAssertEqual(keys?.count, 1)
-      expectation.fulfill()
-    }
-    wait(for: [expectation], timeout: 0)
-  }
-  
-  func testFetchExposureKeysError() {
-    let expectation = self.expectation(description: "a call is made to get the diagnosis keys")
-    let exposureManager = defaultExposureManager(withKeys: false, enAPIVersion: .v1)
-    exposureManager.fetchExposureKeys { (keys, error) in
-      XCTAssertEqual(error?.errorCode, ExposureManagerErrorCode.noExposureKeysFound.rawValue)
-      XCTAssertNotNil(error?.underlyingError)
-      expectation.fulfill()
-    }
-    wait(for: [expectation], timeout: 0)
-  }
-  
-  func testFetchLastDetectionDate() {
-    let userState = UserState()
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1, userState: userState)
-    
-    // Last detection date should initially be nil
-    exposureManager.fetchLastDetectionDate { (posixDate, error) in
-      XCTAssertNil(posixDate)
-      XCTAssertEqual(error?.errorCode, ExposureManagerErrorCode.detectionNeverPerformed.rawValue)
-    }
-    
-    let date = Date()
-    userState.lastExposureCheckDate = date
-    exposureManager.fetchLastDetectionDate { (posixDate, error) in
-      XCTAssertNil(error)
-      XCTAssertEqual(posixDate, NSNumber(value: date.posixRepresentation))
-    }
-  }
-  
-  func testRemainingFileCapacityReset() {
-    let expectation = self.expectation(description: "remainingFileCapacity is updated")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .success:
-        XCTAssertEqual(exposureManager.btSecureStorage.remainingDailyFileProcessingCapacity, 14)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  func testUpdateLastExposureDetectionDateV1() {
-    let expectation = self.expectation(description: "lastExposureCheckDate is updated")
-    
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    
-    XCTAssertNil(exposureManager.btSecureStorage.lastExposureCheckDate)
-    exposureManager.detectExposuresV1 { _ in
-      XCTAssertNotNil(exposureManager.btSecureStorage.lastExposureCheckDate)
-      expectation.fulfill()
-    }
-    wait(for: [expectation], timeout: 50)
-  }
-  
-  func testRemainingFileCapacityNoResetForDetectionError() {
-    let expectation = self.expectation(description: "remainingFileCapacity is not updated")
-    
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1,
-                                                 forceExposureDetectionError: true)
-    
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .failure:
-        XCTAssertEqual(exposureManager.btSecureStorage.userState.remainingDailyFileProcessingCapacity, Constants.dailyFileProcessingCapacity)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  func testDebugFetchDiagnosisKeys() {
-    let debugAction = DebugAction.fetchDiagnosisKeys
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpetactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpetactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpetactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDebugSimulateExposureDetectionError() {
-    let debugAction = DebugAction.simulateExposureDetectionError
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpectactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpectactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpectactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDebugSimulateExposure() {
-    let debugAction = DebugAction.simulateExposure
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpectactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpectactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpectactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDebugFetchExposures() {
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let debugAction = DebugAction.fetchExposures
-    let successExpetactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpetactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpetactionResolve, successExpectationReject], timeout: 2)
-  }
-  
-  func testDebugResetExposures() {
-    let debugAction = DebugAction.resetExposures
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpetactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpetactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpetactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDebugToggleENAuthorization() {
-    let debugAction = DebugAction.toggleENAuthorization
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpetactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpetactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpetactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDebugShowLastProcessedFilePath() {
-    let debugAction = DebugAction.showLastProcessedFilePath
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    let successExpetactionResolve = self.expectation(description: "resolve is called")
-    let successExpectationReject = self.expectation(description: "reject is not called")
-    successExpectationReject.isInverted = true
-    exposureManager.handleDebugAction(debugAction, resolve: { (success) in
-      successExpetactionResolve.fulfill()
-    }) { (_, _, _) in
-      successExpectationReject.fulfill()
-    }
-    wait(for: [successExpetactionResolve, successExpectationReject], timeout: 0)
-  }
-  
-  func testDetectExposuresKeysDownloadingError() {
-    let expectation = self.expectation(description: "the exposure detection call returns an error")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1,
-                                                 forceDownloadKeyError: true)
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .failure(let error):
-        XCTAssertEqual(error.localizedDescription, GenericError.unknown.localizedDescription)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  func testDetectExposuresUnpackagingError() {
-    let expectation = self.expectation(description: "the exposure detection call returns an error")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1,
-                                                 forceKeyUnpackingError: true)
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .failure(let error):
-        XCTAssertEqual(error.localizedDescription, GenericError.unknown.localizedDescription)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  func testDetectExposuresGetExposureInfoError() {
-    let expectation = self.expectation(description: "the exposure detection call returns an error")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1, forceGetExposureInfoError: true)
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .failure(let error):
-        XCTAssertEqual(error.localizedDescription, GenericError.unknown.localizedDescription)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  func testExposureSummaryScoringMatchedKey0() {
-    let enExposureSummary = MockENExposureDetectionSummary()
-    enExposureSummary.matchedKeyCountHandler = {
-      return 0
-    }
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: ExposureConfigurationV1.placeholder))
-  }
-  
-  func testExposureSummaryScoringMatchedKey1() {
-    let enExposureSummary = MockENExposureDetectionSummary()
-    enExposureSummary.matchedKeyCountHandler = {
-      return 1
-    }
-    enExposureSummary.attenuationDurationsHandler = {
-      return [900,0,0]
-    }
-    let configuration = ExposureConfigurationV1.placeholder
-    XCTAssertTrue(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [800,0,0]
-    }
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [0,900,0]
-    }
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [600,600,0]
-    }
-    XCTAssertTrue(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [0,0,1800]
-    }
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: configuration))
-  }
-  
-  func testExposureSummaryScoringMatchedKey3() {
-    let enExposureSummary = MockENExposureDetectionSummary()
-    enExposureSummary.matchedKeyCountHandler = {
-      return 3
-    }
-    enExposureSummary.attenuationDurationsHandler = {
-      return [900,1800,0]
-    }
-    let configuration = ExposureConfigurationV1.placeholder
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [1800,1800,0]
-    }
-    XCTAssertTrue(enExposureSummary.isAboveScoreThreshold(with: configuration))
-  }
-  
-  func testExposureSummaryScoringMatchedKey4() {
-    let enExposureSummary = MockENExposureDetectionSummary()
-    enExposureSummary.matchedKeyCountHandler = {
-      return 4
-    }
-    enExposureSummary.attenuationDurationsHandler = {
-      return [900,1800,0]
-    }
-    let configuration = ExposureConfigurationV1.placeholder
-    XCTAssertFalse(enExposureSummary.isAboveScoreThreshold(with: configuration))
-    enExposureSummary.attenuationDurationsHandler = {
-      return [1800,1800,0]
-    }
-    XCTAssertTrue(enExposureSummary.isAboveScoreThreshold(with: configuration))
-  }
-  
-  func testDetectExposuresSuccessScoreBelow() {
-    let storeExposureExpectation = self.expectation(description: "The exposure does not get stored")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .success(let files):
-        XCTAssertEqual(files, 1)
-        storeExposureExpectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [storeExposureExpectation], timeout:1)
-  }
-  
-  func testDetectExposuresSuccessScoreAbove() {
-    let expectation = self.expectation(description: "1 exposure is detected")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v1)
-    exposureManager.detectExposuresV1 { (result) in
-      switch result {
-      case .success(let files):
-        XCTAssertEqual(files, 1)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
   }
   
   func testRegisterBackgroundTask() {
@@ -946,7 +528,9 @@ class ExposureManagerTests: XCTestCase {
     exposureManager.scheduleBackgroundTaskIfNeeded()
     wait(for: [submitExpectation], timeout: 0)
   }
-  
+
+  // When the request to download the exposure configuration fails
+  // the placeholder configuration is used
   func testGetExposureConfigurationV1FallbackToDefault() {
     let expectation = self.expectation(description: "it falls back to the placeholder exposure configuration")
     let exposureManager = defaultExposureManager(enAPIVersion: .v1, forceDownloadConfigurationError: true)
@@ -959,170 +543,117 @@ class ExposureManagerTests: XCTestCase {
     }
     wait(for: [expectation], timeout: 5)
   }
-  
-  // MARK: == DETECTION EXPOSURE V2 TESTS ==
-  
-  @available(iOS 13.7, *)
-  func testDetectExposuresV2AggregateDetectError() {
-    let expectation = self.expectation(description: "the exposure detection call returns an error")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2,
-                                                 forceExposureDetectionError: true)
-    exposureManager.detectExposuresV2 { (result) in
-      switch result {
-      case .failure(let error):
-        XCTAssertEqual(error.localizedDescription, GenericError.unknown.localizedDescription)
-        expectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  @available(iOS 13.7, *)
-  func testUpdateLastExposureDetectionDateV2() {
-    let expectation = self.expectation(description: "lastExposureCheckDate is updated")
-    
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2)
-    
-    XCTAssertNil(exposureManager.btSecureStorage.lastExposureCheckDate)
-    
-    exposureManager.detectExposuresV2 { _ in
-      XCTAssertNotNil(exposureManager.btSecureStorage.lastExposureCheckDate)
-      expectation.fulfill()
-    }
-    wait(for: [expectation], timeout: 5)
-  }
-  
-  @available(iOS 13.7, *)
-  func testDetectExposuresSuccessPreexistingSavedExposureForDate() {
-    let storeExposureExpectation = self.expectation(description: "The exposure does not get stored")
-    
-    let userState = UserState()
-    userState.exposures.append(Exposure(id: "3",
-                                        date: halloween.posixRepresentation))
-    
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2, userState: userState)
-    
-    (exposureManager.btSecureStorage as! BTSecureStorageMock).storeExposuresHandler = { exposures in
-      XCTAssertEqual(exposures.count, 0)
-      storeExposureExpectation.fulfill()
-    }
-    
-    exposureManager.detectExposuresV2 { _ in }
-    
-    wait(for: [storeExposureExpectation], timeout: 5)
-  }
-  
-  @available(iOS 13.7, *)
-  func testDetectExposuresSuccessNoPreexistingSavedExposureForDate() {
-    let storeExposureExpectation = self.expectation(description: "The exposure is stored")
-    let userState = UserState()
-    userState.exposures.append(Exposure(id: "3",
-                                        date: Date().posixRepresentation))
-    
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2, userState: userState)
-    (exposureManager.btSecureStorage as! BTSecureStorageMock).storeExposuresHandler = { exposures in
-      XCTAssertEqual(exposures.count, 1)
-      storeExposureExpectation.fulfill()
-    }
-    
-    exposureManager.detectExposuresV2 { _ in }
-    
-    wait(for: [storeExposureExpectation], timeout: 5)
-  }
-  
-  @available(iOS 13.7, *)
-  func testValidDailySummariesConfiguration() {
-    let dict: [String: Any] = ["DailySummariesConfig": ["attenuationDurationThresholds": [40,53,60],
-                                                        "attenuationBucketWeights": [1,1,0.5,0],
-                                                        "reportTypeWeights": [1,0,0,0],
-                                                        "reportTypeWhenMissing": 1,
-                                                        "infectiousnessWeights": [1,1],
-                                                        "infectiousnessWhenDaysSinceOnsetMissing": 1,
-                                                        "daysSinceOnsetToInfectiousness": [[-2,1],
-                                                                                           [-1,1],
-                                                                                           [0,1],
-                                                                                           [1,1],
-                                                                                           [2,1],
-                                                                                           [3,1],
-                                                                                           [4,1],
-                                                                                           [5,1],
-                                                                                           [6,1],
-                                                                                           [7,1],
-                                                                                           [8,1],
-                                                                                           [9,1],
-                                                                                           [10,1],
-                                                                                           [11,1],
-                                                                                           [12,1],
-                                                                                           [13,1],
-                                                                                           [14,1]]],
-                               "triggerThresholdWeightedDuration": 15]
-    let jsonData = try! JSONSerialization.data(withJSONObject: dict)
-    let config = DailySummariesConfiguration.create(from: jsonData)!
-    XCTAssertEqual(config.daysSinceOnsetToInfectiousness[NSNumber(value: ENDaysSinceOnsetOfSymptomsUnknown)], 1)
-  }
-  
-  @available(iOS 13.7, *)
-  func testDetectExposuresV2SuccessScoreBelow() {
-    let storeExposureExpectation = self.expectation(description: "The exposure does not gets stored")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2,
-                                                 forceRiskScore: .belowThreshold)
-    exposureManager.detectExposuresV2 { (result) in
-      switch result {
-      case .success(let files):
-        XCTAssertEqual(files, 1)
-        storeExposureExpectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [storeExposureExpectation], timeout: 5)
-  }
-  
-  @available(iOS 13.7, *)
-  func testDetectExposuresV2SuccessScoreAbove() {
-    let storeExposureExpectation = self.expectation(description: "The exposure gets stored")
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2,
-                                                 forceRiskScore: .aboveThreshold)
-    exposureManager.detectExposuresV2 { (result) in
-      switch result {
-      case .success(let files):
-        XCTAssertEqual(files, 1)
-        storeExposureExpectation.fulfill()
-      default: XCTFail()
-      }
-    }
-    wait(for: [storeExposureExpectation], timeout: 5)
-  }
-  
+
+  // When the request to download the exposure configuration fails
+  // the placeholder configuration is used
   @available(iOS 13.7, *)
   func testGetExposureConfigurationV2FallbackToDefault() {
-    let apiClientMock = APIClientMock { (request, requestType) -> (AnyObject) in
-      return Result<String>.success("indexFilePath") as AnyObject
-    }
-    apiClientMock.downloadRequestHander = { (request, requestType) in
-      return Result<DailySummariesConfiguration>.failure(GenericError.unknown)
-    }
-    let exposureManager = defaultExposureManager(enAPIVersion: .v2,
-                                                 forceDownloadConfigurationError: true)
+
+    let expectation = self.expectation(description: "it falls back to the placeholder exposure configuration")
+    let exposureManager = defaultExposureManager(enAPIVersion: .v2, forceDownloadConfigurationError: true)
     do {
       let config = try await(exposureManager.getExposureConfigurationV2())
       XCTAssertEqual(config, DailySummariesConfiguration.placeholder)
+      expectation.fulfill()
     } catch {
       XCTFail()
     }
+    wait(for: [expectation], timeout: 5)
+
   }
-}
 
-var halloween: Date {
-  var components = DateComponents()
-  components.year = 2019
-  components.month = 10
-  components.day = 31
-  let date = Calendar.current.date(from: components)
-  return date!
-}
+  func testEnabledtatus() {
 
-var startOfDay: Date {
-  return Calendar.current.startOfDay(for: Date())
-}
+    let mockENManager = ENManagerMock()
+    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
 
+    mockENManager.exposureNotificationEnabledHandler = {
+      return true
+    }
+    XCTAssertEqual(exposureManager.enabledState, ExposureManager.EnabledState.enabled)
+    mockENManager.exposureNotificationEnabledHandler = {
+      return false
+    }
+    XCTAssertEqual(exposureManager.enabledState, ExposureManager.EnabledState.disabled)
+
+  }
+
+  func testRecentExposures() {
+    let withExposures = defaultStorage()
+    XCTAssertEqual(withExposures.userState.recentExposures.count, 1)
+  }
+
+  func testAuthorizationStatus() {
+
+    let mockENManager = ENManagerMock()
+    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
+
+    mockENManager.authorizationStatusHandler = {
+      return .authorized
+    }
+    XCTAssertEqual(exposureManager.authorizationState,
+                   ExposureManager.AuthorizationState.authorized)
+    mockENManager.authorizationStatusHandler = {
+      return .notAuthorized
+    }
+    XCTAssertEqual(exposureManager.authorizationState,
+                   ExposureManager.AuthorizationState.unauthorized)
+    mockENManager.authorizationStatusHandler = {
+      return .restricted
+    }
+    XCTAssertEqual(exposureManager.authorizationState,
+                   ExposureManager.AuthorizationState.unauthorized)
+    mockENManager.authorizationStatusHandler = {
+      return .unknown
+    }
+    XCTAssertEqual(exposureManager.authorizationState,
+                   ExposureManager.AuthorizationState.unauthorized)
+  }
+
+  func testBluetoothStatus() {
+
+    let mockENManager = ENManagerMock()
+    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
+
+    mockENManager.exposureNotificationStatusHandler = {
+      return .bluetoothOff
+    }
+    XCTAssertFalse(exposureManager.isBluetoothEnabled)
+    mockENManager.exposureNotificationStatusHandler = {
+      return .active
+    }
+    XCTAssertTrue(exposureManager.isBluetoothEnabled)
+    mockENManager.exposureNotificationStatusHandler = {
+      return .disabled
+    }
+    XCTAssertTrue(exposureManager.isBluetoothEnabled)
+    mockENManager.exposureNotificationStatusHandler = {
+      return .paused
+    }
+    XCTAssertTrue(exposureManager.isBluetoothEnabled)
+    mockENManager.exposureNotificationStatusHandler = {
+      return .restricted
+    }
+    XCTAssertTrue(exposureManager.isBluetoothEnabled)
+    mockENManager.exposureNotificationStatusHandler = {
+      return .unknown
+    }
+    XCTAssertTrue(exposureManager.isBluetoothEnabled)
+  }
+
+  func testGetCurrentENPermissionsStatus() {
+    let mockENManager = ENManagerMock()
+    mockENManager.authorizationStatusHandler = {
+      return .authorized
+    }
+    mockENManager.exposureNotificationEnabledHandler = {
+      return true
+    }
+    let exposureManager = ExposureManager(exposureNotificationManager: mockENManager)
+    exposureManager.getCurrentENPermissionsStatus { (authorized, enabled) in
+      XCTAssertEqual(authorized, ExposureManager.AuthorizationState.authorized.rawValue)
+      XCTAssertEqual(enabled, ExposureManager.EnabledState.enabled.rawValue)
+    }
+  }
+
+}
